Create PlaceTable and PlaceHook
PlaceTable component that displays all of the users places as a table. User can click to go to the specific place and also sort by name or date.

PlaceHook should now be used to get all the users place data. It handels the import from firebase and converts the database object to something we can work with in the app.